using tables to declutter
I thought using tables for the phone numbers organized the content a bit, and made it easier to look at. 

I think the page might be easier to navigate if the emails included on this page (lines 131-135) were moved to /emails.html, but i kept them in this version in case you prefer they stay on this page

I added a <style> in the header in order to align the table in the center & align the text in the table to the left. I know an internal stylesheet is not ideal but I'm not sure how to put in a pull req for two files at a time, so i didnt know how to edit the external css, I hope this is okay!